adds routing for pages stations and journeys
